# Sort track parameters before CKF