--- a/io/src/csv/read_measurements.cpp
+++ b/io/src/csv/read_measurements.cpp
@@ -19,7 +19,7 @@
 namespace traccc::io::csv {
 
 void read_measurements(measurement_reader_output& out,
-                       std::string_view filename) {
+                       std::string_view filename, bool do_sort) {
 
     // Construct the measurement reader object.
     auto reader = make_measurement_reader(filename);
@@ -82,8 +82,10 @@ void read_measurements(measurement_reader_output& out,
         result_measurements.push_back(meas);
     }
 
-    std::sort(result_measurements.begin(), result_measurements.end(),
-              measurement_sort_comp());
+    if (do_sort) {
+        std::sort(result_measurements.begin(), result_measurements.end(),
+                  measurement_sort_comp());
+    }
 }
 
 measurement_container_types::host read_measurements_container(