Performance increaseis not very great right now. I will come back to this once we have fast navigation 

```
./bin/traccc_seeding_example_cuda --input_directory=detray_simulation/toy_detector/n_particles_2000/ --check_performance=true --detector_file=detray_json/toy_detector_geometry.json --material_file=detray_json/toy_detector_homogeneous_material.json --event=10 --track_candidates_range=3:10 --constraint-step-size-mm=10000 --run_cpu=1 --skip=0

AFTER
==> Statistics ... 
- read    106829 spacepoints from 26487 modules
- created  (cpu)  114031 seeds
- created (cuda)  114026 seeds
- created  (cpu) 117470 found tracks
- created (cuda) 117481 found tracks
- created  (cpu) 117470 fitted tracks
- created (cuda) 117481 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2711 ms
                Seeding (cuda)  17 ms
                Seeding  (cpu)  606 ms
           Track params (cuda)  16 ms
           Track params  (cpu)  24 ms
 Track finding with CKF (cuda)  492 ms
  Track finding with CKF (cpu)  7828 ms
  Track fitting with KF (cuda)  261 ms
   Track fitting with KF (cpu)  3496 ms
                     Wall time  15474 ms


BEFORE
==> Statistics ... 
- read    106829 spacepoints from 26487 modules
- created  (cpu)  114031 seeds
- created (cuda)  114027 seeds
- created  (cpu) 117470 found tracks
- created (cuda) 117481 found tracks
- created  (cpu) 117470 fitted tracks
- created (cuda) 117481 fitted tracks
==>Elapsed times...
            Hit reading  (cpu)  2742 ms
                Seeding (cuda)  18 ms
                Seeding  (cpu)  601 ms
           Track params (cuda)  3 ms
           Track params  (cpu)  23 ms
 Track finding with CKF (cuda)  511 ms
  Track finding with CKF (cpu)  7982 ms
  Track fitting with KF (cuda)  272 ms
   Track fitting with KF (cpu)  3560 ms
                     Wall time  15737 ms
```